// src/app/shared/enums/app-alert-messages.enum.ts
export enum AppAlertMessages {
    TITLE_SUCCESS = "Success !",
    TITLE_ERROR = "Error !",

    PASSWORD_SET_SUCCESS = "Password set successfully, you can login using your new password.",
    LOGIN_SUCCESS = "Welcome to yCare.",
    SAVE_SUCCESS = "Record saved successfully.",
    INSERT_SUCCESS = "Record inserted successfully.",
    UPDATE_SUCCESS = "Record modified successfully.",
    DELETE_SUCCESS = "Record deleted successfully.",
    ACTION_FAILED = "Action failed, please try again.",
    RESET_SUCCESS = "Record reset successfully.",
    
    FILE_UPLOAD_SUCCESS = "File uploaded successfully.",
    
    CASE_CREATED_SUCCESS = "Case created successfully.",

}
  